Refactor TestCaseExtractor, JSTestExtractor and JavaTestExtractor
